+++ src/pysm3/__init__.py
@@ -18,3 +18,11 @@ from .utils import (
     apply_smoothing_and_coord_transform,
     map2alm,
 )
+
+from .bandpass_sampler import (  # noqa: E402
+    bandpass_distribution_function,
+    bandpass_kresampling,
+    compute_moments,
+    resample_bandpass,
+    search_optimal_kernel_bandwidth,
+)